requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pymongo>=4.10.1
python-dotenv>=1.0.1
pydantic>=2.10.0
pydantic-settings>=2.6.0
google-genai>=0.2.0
python-multipart>=0.0.12
email-validator>=2.2.0
reportlab>=4.0.0
svglib>=1.5.1
pillow>=10.0.0pyjwt
resend
